move errs package
